+++ docs/docusaurus.config.js
@@ -12,14 +12,13 @@ const versions = require("./versions.json");
 module.exports = {
     title: "Evolve SDK (JVM)",
     tagline: "",
-    url: "https://zepben.github.io/evolve/docs/jvm-sdk/",
+    url: "https://zepben.github.io/",
     baseUrl: "/evolve/docs/jvm-sdk/",
-    onBrokenLinks: "warn",
+    onBrokenLinks: "throw",
     favicon: "img/favicon.ico",
     organizationName: "zepben",
     projectName: "evolve-sdk-jvm",
     themeConfig: {
-        ...zepbenDocusaurusPreset.defaultThemeConfig,
         colorMode: {
             defaultMode: "light",
             disableSwitch: false,
@@ -40,7 +39,7 @@ module.exports = {
                 },
                 {
                     to: "/",
-                    activeBasePath: "docs",
+                    activeBasePath: "/",
                     label: "Docs",
                     position: "left",
                 },
@@ -76,12 +75,13 @@ module.exports = {
                 facetFilters: ["project:jvm-sdk"]
             }
         },
-        metadatas: [{name: 'docsearch:project', content: 'jvm-sdk'}],
+        metadata: [{name: 'docsearch:project', content: 'jvm-sdk'}],
     },
     presets: [
         [
             "@zepben/docusaurus-preset",
             {
+                gtag: { ...zepbenDocusaurusPreset.defaultThemeConfig.gtag },
                 docs: {
                     routeBasePath: '/',
                     sidebarPath: require.resolve("./sidebars.js"),